Added Hiding Mechanic and breaker
Hiding mechanic: created script that hides the player when object is interacted with. for now it teleports player inside the object and sets character to not be able to move except to get out. This also zooms in the camera to limit player vision.
Breaker system starter object done but still need to make gameobjects for lights for it to be turned on or off.